feat: 모바일팀의 요구사항 대로 baseOnBalls, Innings의 타입 float로 변경

+++ be/server/src/main/java/com/baseball/dto/PitcherDto.java
@@ -7,8 +7,8 @@ public class PitcherDto {
     private final String name;
     private final Integer numberOfPitching;
     private final Integer hit;
-    private final Integer baseOnBalls;
-    private final Integer innings;
+    private final Float baseOnBalls;
+    private final Float innings;
 
     private PitcherDto(Builder builder) {
         this.name = builder.name;
@@ -22,8 +22,8 @@ public class PitcherDto {
         private String name;
         private Integer numberOfPitching;
         private Integer hit;
-        private Integer baseOnBalls;
-        private Integer innings;
+        private Float baseOnBalls;
+        private Float innings;
 
         private Builder name(String name) {
             this.name = name;
@@ -40,12 +40,12 @@ public class PitcherDto {
             return this;
         }
 
-        private Builder baseOnBalls(Integer baseOnBalls) {
+        private Builder baseOnBalls(Float baseOnBalls) {
             this.baseOnBalls = baseOnBalls;
             return this;
         }
 
-        private Builder innings(Integer innings) {
+        private Builder innings(Float innings) {
             this.innings = innings;
             return this;
         }
@@ -67,11 +67,11 @@ public class PitcherDto {
         return hit;
     }
 
-    public Integer getBaseOnBalls() {
+    public Float getBaseOnBalls() {
         return baseOnBalls;
     }
 
-    public Integer getInnings() {
+    public Float getInnings() {
         return innings;
     }
